Crate twizzler_display

This crate provides basic interfaces for working with the compositor and display manager.

Currently, you can create and manage windows, and fill their framebuffers, and flip that to the compositor. This interface is unstable, and will be changing.

Structs

Buffer

BufferMut

BufferObject

An object holding a double-buffered compositor surface for a window.

DisplayBuffer

DisplayBufferBase

Rect

WindowConfig

WindowHandle

A window handle for the compositor. Provides a compositing buffer.

Functions

get_display_info

Get the current display information. Current resolution is returned in the width and height fields.